feat(codegen): simplify ts-morph type imports
