+++ packages/codegen/package.json
@@ -21,8 +21,8 @@
       "default": "./dist/src/ts-morph-utils/index.js"
     },
     "./ts-morph/*": {
-      "types": "./dist/src/ts-morph-utils/index.d.ts",
-      "default": "./dist/src/ts-morph-utils/index.js"
+      "types": "./dist/src/ts-morph-utils/*/index.d.ts",
+      "default": "./dist/src/ts-morph-utils/*/index.js"
     }
   },
   "scripts": {
