+++ packages/codegen/src/ts-morph-utils/name-builders/index.ts
@@ -1,0 +1,1 @@
+export * from "./get-variable-names.js";
