+++ packages/codegen/src/ts-morph-utils/templates/document-model/hooks.ts
@@ -1,5 +1,5 @@
+import type { DocumentModelTemplateInputs } from "@powerhousedao/codegen/ts-morph";
 import { ts } from "@tmpl/core";
-import type { DocumentModelTemplateInputs } from "../../name-builders/types.js";
 
 export const documentModelHooksFileTemplate = (
   v: DocumentModelTemplateInputs,
