+++ packages/codegen/src/ts-morph-utils/templates/document-model/src/tests/document-model.test.ts
@@ -1,5 +1,5 @@
+import type { DocumentModelTemplateInputs } from "@powerhousedao/codegen/ts-morph";
 import { ts } from "@tmpl/core";
-import type { DocumentModelTemplateInputs } from "../../../../name-builders/types.js";
 
 export const documentModelTestFileTemplate = (v: DocumentModelTemplateInputs) =>
   ts`
